requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djoser==2.2.0
celery==5.3.4
django-sslserver==0.22
drf-yasg==1.21.7
drf-social-oauth2==2.1.3
oauth2_provider==0.0

# IF YOU ARE INSTALLING A PACKAGE OR LIBRARY, PLEASE BE RESPECTIVE OF OTHERS AND ADD IT HERE AND THE VERSION !
# DON'T MAKE RUBBISH AND CONFUSING SOMEONE ELSE 👈(ﾟヮﾟ👈)

Scrapy==2.10.1
scrapy-user-agents==0.1.1
Twisted==22.10.0 # IGNORE UPDATES
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
redis==5.0.0
django-celery-beat==2.5.0
markdown==3.4.4
django-filter==23.2
gevent==23.9.0.post1